INTPYTHON-771 Absorb integration tests into per-class test files
Merge MongoModelSerializer create/update/round-trip DB tests into
test_mongo_model_serializer.py and polymorphic read tests into
test_polymorphic_embedded_model_serializer.py. Remove the separate
integration test files.
